+++ src/methods/ESPRMUser/index.ts
@@ -54,3 +54,4 @@ import "./AddDaylightBasedAutomation";
 import "./GetAutomations";
 import "./GetAutomationDetail";
 import "./SearchESPBLEDevices";
+import "./AssumeRole";